Update start_emulators.sh
In this version of the code, the variable var is set to 3, which means that the loop will run 3 times. For each iteration of the loop, the script increments the port numbers by 2, starts a Docker container using the specified image and port numbers, waits 10 seconds for the container to start up, and then connects to the Android emulator using the new port number.